refactor(agent): amount of bonuses is float64

+++ internal/agent/internal/dto/dto.go
@@ -13,7 +13,7 @@ const (
 )
 
 type AccrualInfo struct {
-	Order   string `json:"order"`
-	Status  string `json:"status"`
-	Accrual int    `json:"accrual,omitempty"`
+	Order   string  `json:"order"`
+	Status  string  `json:"status"`
+	Accrual float64 `json:"accrual,omitempty"`
 }
